Move tests out of unit directory
* salesforcedx-vscode-apex-replay-debugger package
